chore: move serde_json::Value helper methods to agent_shared
