+++ agent_shared/src/lib.rs
@@ -5,6 +5,7 @@ pub mod error;
 pub mod generic_query;
 pub mod handlers;
 pub mod profile;
+pub mod serde_json_value_ext;
 pub mod url_utils;
 
 pub use ::config::ConfigError;
